refactor: one-type-per-file split + formatting cleanup; land Apocalypse level-object parity
Repo-wide automated code-cleanup pass, verified behavior-preserving
(main build 0/0, test build 0/0, full corpus suite 1177/0; a 134-file
adversarial diff review flagged zero behavioral changes):

- Split multi-type files into per-type files (project "one class per file"
  convention): 38 new files, 1 deletion. e.g. DiscSectorSource.cs ->
  IDiscSectorSource / IsoSectorSource / RawSectorSource.
- Removed unused usings, simplified namespace qualifiers, dropped redundant
  named-argument labels, normalized whitespace/indentation, moved nested
  types to end-of-class, converted expression bodies <-> block bodies.
- Restored inner-for indentation in Vid1MotionCompTests to clear 4 new
  S3973 misleading-indentation analyzer warnings the pass introduced.

Also lands the previously-uncommitted Apocalypse level-object feature
(the only behavioral change in this tree; completed/tested earlier):

- Cross-game level-object placement now covers Apocalypse's full
  bank + PLATFORM-overlay + POWERUP-pickup stack. ApplyTriggerOverlay is
  enabled for Apocalypse in MeshCompanionResolver (was bank-layer-only).
- New Apocalypse pickupType -> items.psx table in PsxPowerupPlacementResolver,
  signature-matched from apocalypse_final.exe (CPowerUp ctor jump table
  @0x800A11EC), chosen via the new ApocalypseItemsMarker discriminator in
  SelectTable.
- CLAUDE.md and docs/backlog updated to match.

+++ src/NeversoftMultitool/Core/Formats/Animation/SkaThawParser.cs
@@ -43,6 +43,7 @@ internal static class SkaThawParser
     //         own timestamp bits are unused (sign stays in the header).
     // T keys are THUG-classic on all variants (global T table, /32 scale).
     private const uint FlagThawHiResTimestamps = 1u << 8;
+
     private const uint FlagThawCompactKeys = 1u << 15;
     // bit 16 = scalar-table byte-width components (always set in-corpus; the
     // decoders infer per-component width from the header, so it isn't gated on)
@@ -124,7 +125,8 @@ internal static class SkaThawParser
         for (var bone = 0; bone < numBones; bone++)
         {
             var rotKeys = DecodeThawQKeys(data, ref qOff, qOff + qSizes[bone], compact, hiResTs, compressTable);
-            var transKeys = SkaCompressedKeyDecoders.DecodeCompressedTKeys(data, ref tOff, tOff + tSizes[bone], compressTable);
+            var transKeys =
+                SkaCompressedKeyDecoders.DecodeCompressedTKeys(data, ref tOff, tOff + tSizes[bone], compressTable);
             tracks[bone] = new SkaBoneTrack
             {
                 BoneIndex = bone,
@@ -242,7 +244,7 @@ internal static class SkaThawParser
         short value;
         if (narrow)
         {
-            value = table?.Q48[data[off]].Scalar ?? (short)0;
+            value = table?.Q48[data[off]].Scalar ?? 0;
             off += 1;
         }
         else if (compact)
